feat(primerlab): update with new rust changes and functions

+++ homeworks/week1/primerlab_ref/src/main.rs
@@ -16,7 +16,7 @@ fn main() {
             .read_line(&mut input)
             .expect("Failed to read line");
 
-        let input: usize = match input.trim().parse() {
+        let input: u32 = match input.trim().parse() {
             Ok(num) => num,
             Err(_) => continue,
         };
